backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
python-dotenv
pydantic[email]
python-jose[cryptography]
google-auth 
requests